# Speed-up hot path of GFN2 geometry optimization 