--- a/src/disp/dftd4.F90
+++ b/src/disp/dftd4.F90
@@ -2032,7 +2032,7 @@ subroutine atm_gradient_latp &
    real(wp), intent(inout) :: sigma(:, :)
    real(wp), intent(inout) :: dEdcn(:)
 
-   integer :: iat, jat, kat, ati, atj, atk, jtr, ktr
+   integer :: iat, jat, kat, nat, ati, atj, atk, jtr, ktr
    real(wp) :: cutoff2
    real(wp) :: rij(3), rjk(3), rik(3), r2ij, r2jk, r2ik
    real(wp) :: c6ij, c6jk, c6ik, cij, cjk, cik, scale
@@ -2040,14 +2040,15 @@ subroutine atm_gradient_latp &
    real(wp), parameter :: sr = 4.0_wp/3.0_wp
 
    cutoff2 = cutoff**2
+   nat = len(mol) ! workaround for legacy Intel Fortran compilers
 
    !$omp parallel do default(none) reduction(+:energies, gradient, sigma, dEdcn) &
-   !$omp shared(mol, r4r2, par, trans, cutoff2, c6, dc6dcn) &
+   !$omp shared(mol, r4r2, par, trans, cutoff2, c6, dc6dcn, nat) &
    !$omp private(iat, ati, jat, atj, kat, atk, c6ij, cij, c6ik, c6jk, cik, cjk, &
    !$omp& rij, r2ij, ktr, rik, r2ik, rjk, r2jk, scale, dE, dG, dS, dCN) &
    !$omp collapse(2) schedule(dynamic,32)
-   do iat = 1, len(mol)
-      do jat = 1, len(mol)
+   do iat = 1, nat
+      do jat = 1, nat
          if (jat > iat) cycle
          ati = mol%at(iat)
          atj = mol%at(jat)