@foxtran 2025-02-05
I have no idea why it does not work with ifort. Try to debug tomorrow