@foxtran 2025-02-05
@Albkat, do not try it to restart: it will still fail. Most probably it is a bug in xtb code with undefined variables. I have a tool for debugging compiler :-)