@Albkat 2025-02-05
Thanks, I'll test this.

> @Albkat, @awvwgk, that is not funny that about 1/20 of all CI runs are failing :(

Regarding CI, there are some stochastic failures with PBC, GFN-FF, and CPCM-X that we are trying to identify. It is still a work in progress.


> @Albkat, do not try it to restart: it will still fail. Most probably it is a bug in xtb code with undefined variables. I have a tool for debugging compiler :-)

Yeah, just double-checking if it is compiler-specific :)