@foxtran 2025-02-05
> Regarding CI, there are some stochastic failures with PBC, GFN-FF, and CPCM-X that we are trying to identify. It is still a work in progress.

They are reproducible. I did it for couple of them. :-)