@foxtran 2025-02-06
@Albkat, it works now for legacy Intel Compilers :-)