@foxtran 2025-02-10
> It would be beneficial to understand what you did there, maybe we could use this in other places too...

I just started to use `collapse(2)` OpenMP statement :-)

Note, most of Fortran compilers does not support triangle cycles therefore one needs to give them rectangular grids and one need to be careful about which elements are skipped for better performance and which scheduler should be used.