@foxtran 2025-02-11
@Albkat, rebased on current master branch