@marvinfriede 2025-02-11
> I just started to use `collapse(2)` OpenMP statement :-)
> 
> Note, most of Fortran compilers does not support triangle cycles therefore one needs to give them rectangular grids and one need to be careful about which elements are skipped for better performance and which scheduler should be used.

It seems the rules on collapse are rather strict (see [here](https://www.ibm.com/docs/en/xffbg/121.141?topic=clauses-collapse)).

I did some testing with the D4 (standalone) code and did not find any improvements: triangular iteration space with only outer loop parallel and rectangular iteration space with collapse(2) yielded the same timings. Just wanted to say that this requires testing and does not always help (but nice find here in xtb) :)