@marvinfriede 2025-02-11
> > triangular iteration space with only outer loop parallel and rectangular iteration space with collapse(2) yielded the same timings.
> 
> How many atoms does test have?

I think I tested 500 and 1000 atoms, maybe also 2000.