@foxtran 2025-02-11
Hmm... That is interesting. I'm using gfortran for testing.