@marvinfriede 2025-02-11
> Hmm... That is interesting. I'm using gfortran for testing.

Same. I will take a look at my test setup again later. Maybe I missed something.