@toxtran 2025-02-11
BTW, do you check cpu time or real time? CPU time should not be changing significantly