@foxtran 2025-02-11
My system with 528 atoms, build from c8c187c98742bc9d7d9df25705c7fcfc6715dc79:
```
Cycle 1:
     SCC iter.                  ...        0 min,  1.490 sec
     gradient                   ...        0 min,  3.551 sec
Cycle 2:
     SCC iter.                  ...        0 min,  3.293 sec
     gradient                   ...        0 min,  4.337 sec
     
 total:
 * wall-time:     0 d,  0 h,  0 min, 55.900 sec
 *  cpu-time:     0 d,  0 h,  4 min, 36.190 sec
 * ratio c/w:     4.941 speedup
 SCF:
 * wall-time:     0 d,  0 h,  0 min, 12.917 sec
 *  cpu-time:     0 d,  0 h,  1 min,  9.732 sec
 * ratio c/w:     5.399 speedup
 ANC optimizer:
 * wall-time:     0 d,  0 h,  0 min, 42.595 sec
 *  cpu-time:     0 d,  0 h,  3 min, 25.240 sec
 * ratio c/w:     4.818 speedup
```

Build from 9c78931790d39f062a84904895ca02a5f5934141:
```
Cycle 1:
     SCC iter.                  ...        0 min,  1.673 sec
     gradient                   ...        0 min,  1.913 sec
Cycle 2:
     SCC iter.                  ...        0 min,  3.203 sec
     gradient                   ...        0 min,  1.524 sec

 total:
 * wall-time:     0 d,  0 h,  0 min, 40.080 sec
 *  cpu-time:     0 d,  0 h,  4 min, 23.575 sec
 * ratio c/w:     6.576 speedup
 SCF:
 * wall-time:     0 d,  0 h,  0 min,  9.735 sec
 *  cpu-time:     0 d,  0 h,  1 min,  5.865 sec
 * ratio c/w:     6.766 speedup
 ANC optimizer:
 * wall-time:     0 d,  0 h,  0 min, 29.952 sec
 *  cpu-time:     0 d,  0 h,  3 min, 16.568 sec
 * ratio c/w:     6.563 speedup
```